Move repository-using walk-in tests into whitelisted test classes
The CustomAccessibility analyzer restricts direct construction of the
repository types to specific classes via [OnlyAccessibleBy]. WalkInTests
was not on that list, failing Backend CI with CACC000. The booking and
availability walk-in tests now live in BookingServiceTests and
AvailabilityServiceTests (both whitelisted); WalkInTests keeps the
WalkInHelper and RestaurantManagementService coverage, which need no
restricted types.
